chatbot/general_chatbot/requirements.txt
flask==3.0.3
python-dotenv==1.0.1
PyPDF2==3.0.1
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.35
langchain-huggingface==0.0.3
flask-cors==4.0.1